가져온 데이터 db저장 기능 추가, 접속 시간 전송여부에 따른 데이터 받아오는 기능 한번만 수행하도록 설정

+++ app/src/main/java/com/example/oneplusone/db/ProductModel.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class ProductModel (
 
     @PrimaryKey(autoGenerate = true)
-    val id: Int? =0,
+    val id: Long? =0,
 
     @ColumnInfo(name="productName")
     var productName: String,
